requirements.txt
Django==1.10.1
django-cors-headers==1.3.1
django-filter==1.0.1
djangorestframework==3.4.7
DukeDSClient==0.2.14
funcsigs==1.0.2
mock==2.0.0
pbr==1.10.0
PyYAML==3.12
requests==2.11.1
six==1.10.0
